Create Terraform module to configure swirrl-ons-datahost project
Issue #33 - Add deploy/terraform directory containing a Terraform
module to configure the swirrl-ons-datahost gcloud project. This
project is used to host deployments of the datahost GraphQL
application.

The module configures firewall rules for the default network, a
service account for CI, and a Workload Identity pool to allow
CircleCI OIDC tokens to be exchanged for a set of temporary GCloud
credentials.