Create Terraform module for the GraphQL application
Issue #33 - Add a graphql Terraform module which defines the
deployment for a VM running a version of the GraphQL application.

The VM runs under a service account with permissions to read from
the datahost-graphql docker repository in the swirrl devops
infrastructure project.

The module defines a variable specifying the digest of the image
to run on the VM.