Deploy current version of the ld-openapi application from CI
Issue #108 - Update the ld-openapi application docker build to
display the SHA of the published image.

Save this image SHA within the container publish job, and use it
to automatically deploy a VM configured with the image within a new
deploy-ldapi-vm CI job. This job configures the dev environment
with the latest image digest and deploys it using a terraform
plan/apply.