Store database on secondary volume
Issue #44 - Attach a secondary data disk to the ldapi server and
update the container configuration to format and attach the volume
at /cache within the container. Update the Docker environment config
to write the database to a file within this directory.
Change db path in docker env config and attach data disk at that path