Include container image digest in instance name
Issue #122 - Add a prefix of the container image digest to the end
of the graphql server instance name. Terraform fails to re-add the
new instance into the load balancer if it has the same name as
before. Since a name change forces a re-creation, the explicit
lifecycle dependency on the null resource is also no longer
necessary.