Save series resources in the triplestore
Add representation of a generic RDF resource. Parse incoming
requests into this representation for the PUT series route.

Format resource property values as strings so they have an un-typed
representation in the output json-ld.

Add time namespace and define Clock protocol for fetching the
current time.

Update the GET and POST routes for series resources to convert
incoming requests to the resource representation and save and load
these from the triplestore instead of duratom.

Add tests for the series routes using a memory store and the
ring handler.