Save revisions to the triplestore
Update db/insert-revision to convert revision requests into RDF and
save them into the triplestore instead of the duratom. Revision ids
are now created per-release instead of globally in the database.

The duratom component is no longer used, so remove references to it
from the integrant config along with the init-key and halt-key
methods.

Specify dcterms:description key when creating series and releases
within round-tripping-revision-test. This prevents 400 responses
when fetching associated revisions.

Only assert on the json keys of the responses instead of testing
the entire structure of the json-ld returned from create routes.

Update the 'fetch release' query to find all revisions linked via
the dh:hasRevision property.