Save and load schemas to/from the triplestore
Update the POST schema route to build a resource representation of
the incoming schema and save it to the triplestore.

Update the GET schema route to load from the triplestore. Note the
output representation of schemas is modified from the raw json-ld
representation.

Update how resources are represented so they can contain multiple
subjects. There is still a requirement for a single root resource
which can refer to child resources. This is required so a schema
resource can reference its columns.