Add ChangeStore protocol for storing change data
Issue #152 - Add a ChangeStore protocol for storing and retrieving
CSV change data. Move the existing implementation into the new
store.triplestore namespace, and add a store.file implementation
which stores change files on the filesystem. Files are stored in
the file store in a location determined by the SHA-1 hash of their
contents.

Update the change and revision handlers to require an implementation
of the change store used to save and retreive change data. The key
returned by the insert-append method is associated with the change
(currently only append) record in the triplestore.

Add a test implementation of the file store in store.temp-file-store
which deletes the file store directory on close. Define an integrant
componemt which deletes the directory in the halt-key! method.

Update base, test and docker configuration files to define a suitable
change store implmeentation for each environment.