Make image publishing optional in the docker build task
Issue #175 - Add an optional :to parameter to the docker build task
which controls how built images are published. Only publish images
to the public swirrl repository if this is set to :remote.

Update the deploy-ld-openapi-service CI task to supply this option
so containers are published.

Add documentation to the docker task.