ldapi: hurl files
I started using Hurl[1] as a sanity check for the API. Might be useful
for others.

You can run it like this:

    bin/hurl-test -t FOO

Note: FOO should be unique on each run, as it's used to name the series.
Repeated runs with the same tag will return unexpected results (for
the hurl script, which expects 201, not 200 etc).

[1]: https://hurl.dev