ldapi: ensure column types match the ones in release schema
Tablecloth guesses the column types based on values present in the
dataset. If we have a column with one row and value 75, in a column
where we expect a floating point value, the column type will be int.

This change updates the `as-dataset` function to optionally take
a `:convert-types` option (along with DataseRow) to do the conversion.