Update application routes to use singular for individual resources
Issue #277 - Update the application routes to use the singular case
to address individual resources e.g. a release now has a path of
/{series-slug}/release/{release-slug}. Routes which return a
collection of resources still use the plural case.

Rename 'change' to 'commit' within revision change routes. This is
the first stage of an upcoming series of changes in terminology from
'change' to 'commit'. The rest of the code still widely refers to
commits as changes to minimise the required changes here.

Update the implementation of the AppUri protocol to reflect the new
URI structure for resources. Rename the change-uri(*) method(s) to
commit-uri(*) and rename the api parameters change-id key to
commit-id within commit-uri*.

Update all tests to use the new URIs.