ldapi: redirect to latest rev. when requesting release as text/csv
The GET RELEASE endpoint now can return 422 when the release has no
revisions (and changes), as there's nothing to redirect to.

Some of the release specific tests were moved to revision test namespace
because of that.

Also: since we use the redirect URL, the rdf-base-uri can't be bogus
anymore (in test context).