Use external context in handler (fixes #348)
The context file jsonld-context.json has been updated to force
dh:hasChange and dh:hasRevision to always be arrays.

That context file is now loaded external so as not to pollute the
@context in the API responses with lots of potentially unnecessary
predicates.

A new `json-ld/context`` function has been created to load the external context
and include the dynamically-generated @base. That function replaces
the `simple-context` and `simple-collection-context` functions.

All handler functions that used `simple-context` or `simple-collection-context`
have been updated to use `json-ld/context`